create add and UI

+++ Model/Product.cs
@@ -9,7 +9,7 @@ namespace AppHondaDuyDuc.Model
     public class Product
     {
         public string Name { get; set; }
-        public float Quantity { get; set; }
-        public float Price { get; set; }
+        public double Quantity { get; set; }
+        public double Price { get; set; }
     }
 }
